Pull out ContextFlowOps from mets_adapter
